cliente_simulador/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.10.0=py38_0
aiohttp=3.6.3=py38h7b6447c_0
apache-beam=2.25.0=pypi_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py38_0
attrs=20.2.0=py_0
avro-python3=1.10.0=pypi_0
blas=1.0=mkl
blinker=1.4=py38_0
brotlipy=0.7.0=py38h7b6447c_1000
c-ares=1.16.1=h7b6447c_0
ca-certificates=2020.10.14=0
cachetools=4.1.1=py_0
certifi=2020.6.20=py38_0
cffi=1.14.3=py38he30daa8_0
chardet=3.0.4=py38_1003
click=7.1.2=py_0
contextlib2=0.6.0.post1=pypi_0
crcmod=1.7=pypi_0
cryptography=3.1.1=py38h1ba5d50_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cupti=10.1.168=0
cycler=0.10.0=pypi_0
cython=0.29.21=pypi_0
dataclasses=0.6=pypi_0
dill=0.3.1.1=pypi_0
dm-tree=0.1.5=pypi_0
docopt=0.6.2=pypi_0
fastavro=1.0.0.post1=pypi_0
future=0.18.2=pypi_0
gast=0.3.3=py_0
gin-config=0.3.0=pypi_0
google-api-core=1.23.0=pypi_0
google-api-python-client=1.12.5=pypi_0
google-auth=1.22.1=py_0
google-auth-httplib2=0.0.4=pypi_0
google-auth-oauthlib=0.4.1=py_2
google-cloud-bigquery=2.2.0=pypi_0
google-cloud-core=1.4.3=pypi_0
google-crc32c=1.0.0=pypi_0
google-pasta=0.2.0=py_0
google-resumable-media=1.1.0=pypi_0
googleapis-common-protos=1.52.0=pypi_0
grpcio=1.31.0=py38hf8bcb03_0
h5py=2.10.0=py38hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
hdfs=2.5.8=pypi_0
httplib2=0.17.4=pypi_0
idna=2.10=py_0
importlib-metadata=2.0.0=py_1
importlib-resources=3.1.1=pypi_0
intel-openmp=2020.2=254
kaggle=1.5.9=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.2.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libprotobuf=3.13.0.1=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
lxml=4.6.1=pypi_0
markdown=3.3.2=py38_0
matplotlib=3.3.2=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
mock=2.0.0=pypi_0
multidict=4.7.6=py38h7b6447c_1
ncurses=6.2=he6710b0_1
numpy=1.19.1=py38hbc911f0_0
numpy-base=1.19.1=py38hfa32c7d_0
oauth2client=4.1.3=pypi_0
oauthlib=3.1.0=py_0
object-detection=0.1=pypi_0
opencv-python=4.4.0.44=pypi_0
opencv-python-headless=4.4.0.44=pypi_0
openssl=1.1.1h=h7b6447c_0
opt_einsum=3.1.0=py_0
pandas=1.1.3=pypi_0
pbr=5.5.1=pypi_0
pillow=8.0.1=pypi_0
pip=20.2.4=py38_0
promise=2.3=pypi_0
proto-plus=1.11.0=pypi_0
protobuf=3.13.0.1=py38he6710b0_1
psutil=5.7.3=pypi_0
py-cpuinfo=7.0.0=pypi_0
pyarrow=0.17.1=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycocotools=2.0.2=pypi_0
pycparser=2.20=py_2
pydot=1.4.1=pypi_0
pyjwt=1.7.1=py38_0
pymongo=3.11.0=pypi_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pypi_0
pysocks=1.7.1=py38_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=pypi_0
python-slugify=4.0.1=pypi_0
pytz=2020.1=pypi_0
pyyaml=5.3.1=pypi_0
readline=8.0=h7b6447c_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
scipy=1.5.2=py38h0b6359f_0
sentencepiece=0.1.94=pypi_0
setuptools=50.3.0=py38hb0f4dca_1
six=1.15.0=py_0
slugify=0.0.1=pypi_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.3.0=pypi_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.3.1=pypi_0
tensorflow-addons=0.11.2=pypi_0
tensorflow-datasets=4.0.1=pypi_0
tensorflow-estimator=2.3.0=pypi_0
tensorflow-gpu=2.2.0=h0d30ee6_0
tensorflow-hub=0.9.0=pypi_0
tensorflow-metadata=0.24.0=pypi_0
tensorflow-model-optimization=0.5.0=pypi_0
termcolor=1.1.0=py38_1
text-unidecode=1.3=pypi_0
tf-models-official=2.3.0=pypi_0
tf-slim=1.1.0=pypi_0
tk=8.6.10=hbc83047_0
tqdm=4.50.2=pypi_0
typeguard=2.10.0=pypi_0
typing-extensions=3.7.4.3=pypi_0
uritemplate=3.0.1=pypi_0
urllib3=1.25.11=py_0
werkzeug=1.0.1=py_0
wheel=0.35.1=py_0
wrapt=1.12.1=py38h7b6447c_1
xz=5.2.5=h7b6447c_0
yarl=1.6.2=py38h7b6447c_0
zipp=3.3.1=py_0
zlib=1.2.11=h7b6447c_3
